Remove hard-coding in student method for application controller

+++ db/migrate/20200707064558_drop_course_session_users.rb
@@ -1,5 +1,5 @@
 class DropCourseSessionUsers < ActiveRecord::Migration[6.0]
   def change
-    drop_table :course_session_users
+    drop_table :course_session_users if table_exists? :course_session_users
   end
 end
